Refactoring lib menu to organization.
